Work on making cross platform with macOS

+++ src/TextRenderer.cpp
@@ -61,7 +61,7 @@ TextRenderer::TextRenderer(glm::mat4 proj) : _textShader("shaders/text_shader.ve
 			texture,
 			glm::ivec2(_face->glyph->bitmap.width, _face->glyph->bitmap.rows),
 			glm::ivec2(_face->glyph->bitmap_left, _face->glyph->bitmap_top),
-			_face->glyph->advance.x
+			static_cast<unsigned int>(_face->glyph->advance.x)
 		};
 		Characters.insert(std::pair<GLchar, Character>(c, character));
 	}
